Change MaxWordSeg SepType::Score type from float to int.
	All calculations are already in int, and the casts are just
	unnecessary, and indicates that "float" is never intended.
	Change short int to int to be more machine-native.

	* src/worddef.h:
	  - Change SepType::Score type to int.
	* src/maxwordseg.cpp (MaxWordSeg::WordSegArea):
	  - Change types of "bestScore", "tmps", "score[]" to int.
	  - Drop all (short int) type casts.

+++ src/worddef.h
@@ -37,7 +37,7 @@ typedef struct CDAreaDef
 
 typedef struct SepTypeDef
 {
-  float Score;
+  int Score;
   short int Sep[MAXSEP];
 } SepType;
